projects/advanced/chat-with-pdf/requirements.txt
google-genai
python-dotenv
langchain
langchain-google-genai
langchain-community
pypdf
faiss-cpu